Update Project
I've Add UtilsVeicoliDLLProject file , which it includes all the Classes named by Utils, choosen by me:
Utils.Database(all function to manipulate a connection between the project and the Database, with all command to execute SQL, and Database).
Utils.Word(all the function for OpenXml-Word to create a page .docx)
Utils.Excel(all the function for Openxml-Excel to create a page .xlsx).
All these Classes are included in that file.
I've Added some code to optimize and to advance the Car-Shop-project .